+++ lib/src/utils/date_picker_controller.dart
@@ -2,26 +2,17 @@ import 'package:flutter/cupertino.dart';
 
 class DatePickerController extends FixedExtentScrollController {
   DatePickerController({
-    this.year = 2021,
-    this.month = 1,
-    this.day = 1,
+    required this.initialDateTime,
     this.minYear = 2010,
     this.maxYear = 2050,
-  })  : monthController = FixedExtentScrollController(initialItem: month - 1),
-        dayController = FixedExtentScrollController(initialItem: day - 1),
-        assert(year >= minYear),
+  })  : monthController =
+            FixedExtentScrollController(initialItem: initialDateTime.month - 1),
+        dayController =
+            FixedExtentScrollController(initialItem: initialDateTime.day - 1),
+        assert(initialDateTime.year >= minYear),
         super(
-          initialItem: year - minYear,
+          initialItem: initialDateTime.year - minYear,
         );
-
-  /// The initial date and/or time of the picker.
-  final int year;
-
-  /// The initial date and/or time of the picker.
-  final int month;
-
-  /// The initial date and/or time of the picker.
-  final int day;
 
   /// Minimum year that the picker can be scrolled
   final int minYear;
@@ -34,6 +25,9 @@ class DatePickerController extends FixedExtentScrollController {
 
   /// This widget's day selection and animation state.
   final FixedExtentScrollController dayController;
+
+  /// The initial date and/or time of the picker.
+  final DateTime initialDateTime;
 
   @override
   void dispose() {